refactor : remove legacy mvc and refactor using annotation
