+++ app/src/main/java/camp/nextstep/controller/IndexController.java
@@ -1,0 +1,18 @@
+package camp.nextstep.controller;
+
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.view.JspView;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Controller
+public class IndexController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView index(final HttpServletRequest req, final HttpServletResponse resp) {
+        return new ModelAndView(new JspView("/index.jsp"));
+    }
+}
